requirements.txt
absl-py==2.0.0
accelerate==0.24.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
basicsr==1.4.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
blobfile==2.1.1
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2022.12.7
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
config==0.5.1
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.14.6
decord==0.6.0
diffusers==0.22.0.dev0
dill==0.3.7
distlib==0.3.8
docker-pycreds==0.4.0
easydict==1.11
einops==0.7.0
fairscale==0.4.4
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gradio==4.3.0
gradio_client==0.7.0
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
identify==2.5.35
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kaggle==1.6.6
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lightning-utilities==0.9.0
lmdb==1.4.1
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.0
nodeenv==1.8.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.10
pandas==2.1.2
Pillow==9.3.0
plotly==5.19.0
portalocker==2.8.2
pre-commit==3.6.2
preshed==3.0.9
protobuf==4.23.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycryptodomex==3.19.0
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.8.1b0
pyDeprecate==0.3.2
pydub==0.25.1
pyparsing==3.1.1
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.4
pytorch-lightning==1.7.7
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.31.0
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
salesforce-lavis==1.0.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
share==1.0.4
shellingham==1.5.4
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.0+cu118
torch-tb-profiler==0.4.3
torchaudio==2.1.0+cu118
torchmetrics==1.2.0
torchvision==0.16.0+cu118
tqdm==4.66.1
transformers==4.26.1
triton==2.1.0
typer==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.13
uvicorn==0.24.0.post1
validators==0.22.0
virtualenv==20.25.0
wandb==0.16.0
wasabi==1.1.2
watchdog==4.0.0
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
xformers==0.0.22.post7
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
